test-apps/web-chatbot/pyproject.toml
[project]
name = "web-chatbot"
version = "0.1.0"
description = "Web chatbot with local UI for testing prompt injection"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "litellm",
    "python-dotenv",
    "httpx",
]
